requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
Cython==0.29.32
frozenlist==1.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
idna==3.4
magic-filter==1.0.9
miniful==0.0.6
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
pyFUME==0.2.25
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
scipy==1.10.1
sgmllib3k==1.0.0
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
urllib3==1.26.14
yarl==1.8.2
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
Cython==0.29.32
frozenlist==1.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
miniful==0.0.6
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
pyFUME==0.2.25
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
scipy==1.10.1
sgmllib3k==1.0.0
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
urllib3==1.26.14
yarl==1.8.2
